Fix erroneous calculation of match column width.
The width of the contents of match column cells (i.e., of match labels)
was being artificially increased by 1 for some reason, which threw off
column alignment when the longest match was longer than the longest
filename.

Also write an extra space before the filename row to compensate and
correctly align the filename labels.